VPLAY-11310: Remove legacy AAMP JS bindings
Reason for change: Remove un-used AAMP legacy JS bindings
Test Procedure: Make sure all apps are working as expected
Risks: None
Priority: P2
