--- jsbindings/jsbindings-version.h
@@ -1,35 +1,0 @@
-/*
- * If not stated otherwise in this file or this component's license file the
- * following copyright and licenses apply:
- *
- * Copyright 2020 RDK Management
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-#ifndef __JSBINDINGS_VERSION_H__
-#define __JSBINDINGS_VERSION_H__
-
-#include "AampDefine.h"
-/*
- * @file jsbindings-version.h
- * @brief Holds the AAMP JS bindings version information.
- */
-
-#define AAMP_UNIFIED_VIDEO_ENGINE_VERSION AAMP_VERSION
-
-// GLOBAL_AAMP_NATIVEBINDING_VERSION is no longer used as this will be deprecated in future
-//#define GLOBAL_AAMP_NATIVEBINDING_VERSION "2.7"
-
-#endif /* __JSBINDINGS_VERSION_H__ */
-
