+++ jsbindings/jseventlistener.h
@@ -27,8 +27,7 @@
 #define __AAMP_JSEVENTLISTENER__H__
 
 
-#include "jsbindings.h"
-
+#include "jsbindings-main.h"
 
 /**
  * @class AAMP_JSEventListener
@@ -68,9 +67,9 @@ public:
 	 */
 	~AAMP_JSEventListener();
 	/**
-         * @brief Copy constructor disabled
-         *
-         */
+	 * @brief Copy constructor disabled
+	 *
+	 */
 	AAMP_JSEventListener(const AAMP_JSEventListener&) = delete;
 	/**
  	 * @brief assignment operator disabled
@@ -85,7 +84,7 @@ public:
 
 	/**
 	 * 
-	*/
+	 */
 	virtual void SetEventProperties(const AAMPEventPtr& ev, JSObjectRef jsEventObj);
 
 public:
