+++ main_aamp.cpp
@@ -63,8 +63,6 @@ PlayerInstanceAAMP::PlayerInstanceAAMP(StreamSink* streamSink
 	// tune only . After that every tune will use the same config parameters
 	if(gpGlobalConfig == NULL)
 	{
-		
-
 		curl_global_init(CURL_GLOBAL_DEFAULT);
 		auto vers = curl_version_info(CURLVERSION_NOW);
 		printf( "curl version: %s\n", vers->version );
@@ -99,16 +97,6 @@ PlayerInstanceAAMP::PlayerInstanceAAMP(StreamSink* streamSink
 	std::shared_ptr<PlayerExternalsInterface> pExternalsInterface = PlayerExternalsInterface::GetPlayerExternalsInterfaceInstance();
 	pExternalsInterface->SetUseFireBoltSDK(gpGlobalConfig->IsConfigSet(eAAMPConfig_UseFireboltSDK));
 	pExternalsInterface->Initialize();
-
-#ifdef SUPPORT_JS_EVENTS
-#ifdef AAMP_WPEWEBKIT_JSBINDINGS //aamp_LoadJS defined in libaampjsbindings.so
-	const char* szJSLib = "libaampjsbindings.so";
-#else
-	const char* szJSLib = "libaamp.so";
-#endif
-	mJSBinding_DL = dlopen(szJSLib, RTLD_GLOBAL | RTLD_LAZY);
-	AAMPLOG_WARN("[AAMP_JS] dlopen(\"%s\")=%p", szJSLib, mJSBinding_DL);
-#endif
 
 #ifdef AAMP_BUILD_INFO
 		std::string tmpstr = MACRO_TO_STRING(AAMP_BUILD_INFO);
@@ -205,13 +193,6 @@ PlayerInstanceAAMP::~PlayerInstanceAAMP()
 	{
 		PlayerCCManager::DestroyInstance();
 	}
-#ifdef SUPPORT_JS_EVENTS
-	if (mJSBinding_DL && isLastPlayerInstance)
-	{
-		AAMPLOG_WARN("[AAMP_JS] dlclose(%p)", mJSBinding_DL);
-		dlclose(mJSBinding_DL);
-	}
-#endif
 	if (isLastPlayerInstance)
 	{
 		ContentSecurityManager::DestroyInstance();
@@ -1541,7 +1522,8 @@ void PlayerInstanceAAMP::SetSubscribedTags(std::vector<std::string> subscribedTa
 		UsingPlayerId playerId(aamp->mPlayerId);
 		aamp->subscribedTags = subscribedTags;
 
-		for (int i=0; i < aamp->subscribedTags.size(); i++) {
+		for (int i=0; i < aamp->subscribedTags.size(); i++)
+		{
 			AAMPLOG_WARN("    subscribedTags[%d] = '%s'", i, subscribedTags.at(i).data());
 		}
 	}
@@ -1563,50 +1545,14 @@ void PlayerInstanceAAMP::SubscribeResponseHeaders(std::vector<std::string> respo
 	}
 }
 
-#ifdef SUPPORT_JS_EVENTS
-
-/**
- *  @brief Load AAMP JS object in the specified JS context.
- */
-void PlayerInstanceAAMP::LoadJS(void* context)
-{
-	AAMPLOG_WARN("[AAMP_JS] (%p)", context);
-	if (mJSBinding_DL) {
-		void(*loadJS)(void*, void*);
-		const char* szLoadJS = "aamp_LoadJS";
-		loadJS = (void(*)(void*, void*))dlsym(mJSBinding_DL, szLoadJS);
-		if (loadJS) {
-			AAMPLOG_WARN("[AAMP_JS]  dlsym(%p, \"%s\")=%p", mJSBinding_DL, szLoadJS, loadJS);
-			loadJS(context, this);
-		}
-	}
-}
-
-/**
- *  @brief Unload AAMP JS object in the specified JS context.
- */
-void PlayerInstanceAAMP::UnloadJS(void* context)
-{
-	AAMPLOG_WARN("[AAMP_JS] (%p)", context);
-	if (mJSBinding_DL) {
-		void(*unloadJS)(void*);
-		const char* szUnloadJS = "aamp_UnloadJS";
-		unloadJS = (void(*)(void*))dlsym(mJSBinding_DL, szUnloadJS);
-		if (unloadJS) {
-			AAMPLOG_WARN("[AAMP_JS] dlsym(%p, \"%s\")=%p", mJSBinding_DL, szUnloadJS, unloadJS);
-			unloadJS(context);
-		}
-	}
-}
-#endif
-
 /**
  *  @brief Support multiple listeners for multiple event type
  */
 void PlayerInstanceAAMP::AddEventListener(AAMPEventType eventType, EventListener* eventListener)
 {
-	if(aamp){
-	aamp->AddEventListener(eventType, eventListener);
+	if(aamp)
+	{
+		aamp->AddEventListener(eventType, eventListener);
 	}
 }
 
@@ -1615,8 +1561,9 @@ void PlayerInstanceAAMP::AddEventListener(AAMPEventType eventType, EventListener
  */
 void PlayerInstanceAAMP::RemoveEventListener(AAMPEventType eventType, EventListener* eventListener)
 {
-	if(aamp){
-	aamp->RemoveEventListener(eventType, eventListener);
+	if(aamp)
+	{
+		aamp->RemoveEventListener(eventType, eventListener);
 	}
 }
 
@@ -1637,10 +1584,9 @@ bool PlayerInstanceAAMP::IsLive()
  */
 
 bool PlayerInstanceAAMP::IsJsInfoLoggingEnabled(void)
-
- {
-	 return  ISCONFIGSET(eAAMPConfig_JsInfoLogging);
- }
+{
+	return ISCONFIGSET(eAAMPConfig_JsInfoLogging);
+}
 
 /**
  *  @brief Get current audio language.
