+++ test/utests/tests/JsBindingTests/JsBindingTests.cpp
@@ -16,10 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "jsbindings.h"
-#include "jsevent.h"
-#include "jseventlistener.h"
-#include "jsutils.h"
+#include "jsbindings-main.h"
 #include "jsutils.h"
 #include "PersistentWatermark.h"
 #include "PersistentWatermarkDisplaySequencer.h"
@@ -56,9 +53,9 @@ public:
 TEST_F(JsBindingTests, TestJsBindings)
 {
 	void *context = NULL;
-	aamp_LoadJS(context, playerInstanceAAMP);
+	AAMPPlayer_LoadJS(context);
 
-	aamp_UnloadJS( context );
+	AAMPPlayer_UnloadJS(context);
 }
 
 TEST_F(JsBindingTests, TestJsonUtils )
